// rihal/apps/scrap-digikala/src/scraper/model/DigiItemEntity.ts
export interface DigiItemEntity {
  digiId: number;
  title_fa: any;
  title_en: any;
  uri: any;
  brand: any;
  item_category2: any;
  item_category3: any;
  item_category4: any;
  item_category5: any;
  rating_rate: any;
  rating_count: any;
  mainImage: any;
   size: any;
   seller_code: any;
   seller_title: any;
  seller_rating_commitment: any;
  seller_rating_count: any;
  selling_price: any;
  rrp_price: any;
  discount_percent: any;
  //badge_title: any;
  shipment_methods: any;
  created_at: any;
}